// src/fin/types.d.ts
type DenomResponse = { native: string } | { cw20: string } | string;

export type Config = {
  owner: string;
  denoms: [Denom, Denom];
  isBootstrapping: boolean;
  decimalDelta: number;
};

export type Order = {
  idx: number;
  owner: string;
  quotePrice: number;
  offerDenom: Denom;
  offerAmount: BigNumber;
  filledAmount: BigNumber;
  createdAt: Date;
  originalOfferAmount: BigNumber;
};

export type Book = {
  base: Pool[];
  quote: Pool[];
};

export type Margin = {
  market: string;
  limit?: string;
  fin_address: string;
  config: Market;
};

export type Pair = {
  address: string;
  chainID: NETWORK;
  denoms: [Denom, Denom];
  precision: Precision;
  decimalDelta: number;
  margin?: Margin;
  multiswap: boolean;
  pool?: string;
  queue?: string;
  calc?: string;
};

export type Pool = {
  quotePrice: number;
  offerDenom: Denom;
  totalOfferAmount: BigNumber;
};

export type Simulation = {
  returnAmount: BigNumber;
  spreadAmount: BigNumber;
  commissionAmount: BigNumber;
  rate: number;
  slippage: number;
};

export type CW20Denom =
  | {
      native: string;
    }
  | {
      cw20: Addr;
    };
export type Addr = string;
export type Decimal = string;
export type Uint128 = string;
export interface BookResponse {
  base: PoolResponse[];
  quote: PoolResponse[];
}
export interface PoolResponse {
  offer_denom: CW20Denom;
  quote_price: Decimal;
  total_offer_amount: Uint128;
}
export type Precision =
  | { significant_figures: number }
  | { decimal_places: number };

export interface ConfigResponse {
  denoms: [CW20Denom, CW20Denom];
  is_bootstrapping: boolean;
  owner: Addr;
  price_precision: Precision;
  decimal_delta: number;
}
// ...
export type Timestamp = Uint64;
export type Uint64 = string;
export interface OrderResponse {
  created_at: Timestamp;
  filled_amount: Uint128;
  idx: Uint128;
  offer_amount: Uint128;
  offer_denom: CW20Denom;
  original_offer_amount: Uint128;
  owner: Addr;
  quote_price: Decimal;
}